+++ cmake/nfxGraphicsSources.cmake
@@ -5,5 +5,6 @@
 set(NFX_GRAPHICS_SOURCES
     src/gl/mesh/Mesh.cpp
     src/gl/mesh/Primitive.cpp
+    src/gl/pipeline/RenderState.cpp
     src/gl/pipeline/RenderTarget.cpp
 )